# Add progress output