--- a/books/RayTracingInOneWeekend.html
+++ b/books/RayTracingInOneWeekend.html
@@ -105,9 +105,9 @@ There are some things to note in that code:
      fully on at the top. Red and green together make yellow so we should expect the upper right
      corner to be yellow.
 
-Because the file is written to the program output, you'll need to capture the program output.
-Typically this is done from the command-line by redirecting output into a file (using the `>`
-redirection operator), like so:
+<div class='together'>
+Because the file is written to the program output, you'll need to redirect it to the image file.
+Typically this is done from the command-line by using the `>` redirection operator, like so:
 
     ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
     build\Release\inOneWeekend.exe > image.ppm
@@ -118,6 +118,7 @@ This is how things would look on Windows. On Mac or Linux, it would look like th
     ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
     build/Release/inOneWeekend > image.ppm
     ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+</div>
 
 <div class='together'>
 Opening the output file (in `ToyViewer` on my Mac, but try it in your favorite viewer and Google