--- a/books/RayTracingInOneWeekend.html
+++ b/books/RayTracingInOneWeekend.html
@@ -106,7 +106,7 @@ There are some things to note in that code:
      corner to be yellow.
 
 <div class='together'>
-Because the file is written to the program output, you'll need to redirect it to the image file.
+Because the file is written to the program output, you'll need to redirect it to an image file.
 Typically this is done from the command-line by using the `>` redirection operator, like so:
 
     ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
@@ -116,7 +116,7 @@ Typically this is done from the command-line by using the `>` redirection operat
 This is how things would look on Windows. On Mac or Linux, it would look like this:
 
     ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-    build/Release/inOneWeekend > image.ppm
+    build/inOneWeekend > image.ppm
     ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 </div>