--- a/src/InOneWeekend/main.cc
+++ b/src/InOneWeekend/main.cc
@@ -95,10 +95,11 @@ int main() {
 
     camera cam(lookfrom, lookat, vec3(0,1,0), 20, double(nx)/ny, aperture, dist_to_focus);
 
-    for (int j = ny-1; j >= 0; j--) {
-        for (int i = 0; i < nx; i++) {
+    for (int j = ny-1; j >= 0; --j) {
+        std::cerr << "\rScanlines remaining: " << j << ' ' << std::flush;
+        for (int i = 0; i < nx; ++i) {
             vec3 color;
-            for (int s = 0; s < num_samples; s++) {
+            for (int s = 0; s < num_samples; ++s) {
                 auto u = double(i + random_double()) / double(nx);
                 auto v = double(j + random_double()) / double(ny);
                 ray r = cam.get_ray(u, v);
@@ -107,4 +108,6 @@ int main() {
             color.write_color(std::cout, num_samples);
         }
     }
+
+    std::cerr << "\nDone.\n";
 }