--- a/src/TheRestOfYourLife/main.cc
+++ b/src/TheRestOfYourLife/main.cc
@@ -99,10 +99,11 @@ int main() {
     a[1] = glass_sphere;
     hittable_list hlist(a,2);
 
-    for (int j = ny-1; j >= 0; j--) {
-        for (int i = 0; i < nx; i++) {
+    for (int j = ny-1; j >= 0; --j) {
+        std::cerr << "\rScanlines remaining: " << j << ' ' << std::flush;
+        for (int i = 0; i < nx; ++i) {
             vec3 color;
-            for (int s = 0; s < num_samples; s++) {
+            for (int s = 0; s < num_samples; ++s) {
                 auto u = (i + random_double()) / nx;
                 auto v = (j + random_double()) / ny;
                 ray r = cam->get_ray(u, v);
@@ -112,4 +113,6 @@ int main() {
             color.write_color(std::cout, num_samples);
         }
     }
+
+    std::cerr << "\nDone.\n";
 }