Resolves #139