@trevordblack 2019-11-25
Wow! I really like this change.

Should we propagate it into the text as well?
This solution will need slight tweaking when we multithread the code base.

I benchmarked it, and it looks indistinguishable from before

Without change:
```
% time ./inOneWeekend > inoneweekend.ppm

Done.
./inOneWeekend > inoneweekend.ppm  294.00s user 0.15s system 99% cpu 4:54.23 total
```

With Change:
```
% time ./inOneWeekend > inoneweekend.ppm
Scanlines remaining: 0   
Done.
./inOneWeekend > inoneweekend.ppm  294.11s user 0.30s system 99% cpu 4:54.72 total
```