@hollasch 2019-11-25
Gah! I remembered and then spaced on synchronizing with the text. I'll add that later, so we'll sit on this pull until that's done.