@hollasch 2019-11-25
@trevordblack — let me know what you think about the text changes.