@trevordblack 2019-11-25
The addition of "Adding a Progress Indicator" adds a chapter 2.1

No other chapters in _InOneWeekend_ have subchapters.

The only chapters that have subchapters are in _TheNextWeek_ and _TheRestOfYourLife_.
These subchapters are for those books' Acknowledgments subchapter.

We either need to remove the 2.1 for Progress Indicator, or start partitioning other chapters.