@trevordblack 2019-11-25
Changes made in `Incorporate review feedback` solve 2/3 problems. See third commit above.