@hollasch 2019-11-25
Subchapters would definitely help in many cases. It's just another tool we have available for organization. Feel free to introduces subchapter headers as needed.