@trevordblack 2019-11-25
Subchapters would be a hugely significant change.

I'm opening another issue to hash out that discussion separately.

As least for now--as far as this branch goes--I approve of all changes made.
And approve the pull request.